% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetagene.R
\name{addSym}
\alias{addSym}
\title{Provide gene symbols for transcripts in metaList}
\usage{
addSym(
  metaList = NULL,
  bed = NULL,
  covList = NULL,
  cols = NULL,
  species = "hsapiens",
  rmZeroes = T,
  control = "scr_total"
)
}
\arguments{
\item{metaList}{object created from makeCovList(). Used by default over 'bed' and 'covList' if provided.}

\item{bed}{data.frame of regions associated with covList. Leave NULL if providing metaList}

\item{covList}{list of data.frame coverage matrices. Leave NULL if providing metaList}

\item{cols}{numeric vector of suggested columns for plotMetaGene() and subsequent functions. leave NULL if providing metaList}

\item{species}{character indicating species from which to derive gene symbols. At the moment only "hsapiens" or "mmusculus" are accepted.}

\item{rmZeroes}{Boolean indicating if rows with no coverage are to be removed from coverage matrices. Default=TRUE.}

\item{control}{character indicating name of control entry from sampleTable given to makeCovList().}
}
\value{
Same list object as returned by makeCovList(), but now gene symbols are annotated in $bed object
}
\description{
Provide gene symbols for transcripts in metaList
}
