% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrAcc.R
\name{aggrAcc}
\alias{aggrAcc}
\title{Aggregate peak counts to gene features}
\usage{
aggrAcc(
  counts,
  anno,
  annoCol = "V4",
  feature = "all",
  method = c("mean", "median", "geoMean", "max", "sum"),
  raw = F,
  geneCol = "SYMBOL"
)
}
\arguments{
\item{counts}{count matrix where rows=peaks and columns=individual samples}

\item{anno}{csAnno object (from ChIPseeker annotatePeak()) containing the annotations for only and all of the peaks in 'counts'}

\item{annoCol}{column name in anno@anno corresponding to the rownames of counts. Defaults to "V4".}

\item{feature}{Type of feature to aggregate. 'all" will be all features. Other options include "promoter","intron","intergenic", etc. Default is "all".}

\item{method}{method to use for aggregating counts. One of "mean","median","geoMean", "max", or "sum". Defaults to "mean"}

\item{raw}{Boolean indicating if exported object should be whole integers (i.e. raw counts). Default=FALSE.}
}
\value{
Data.frame with rows=gene symbols, columns= samples, and aggregated gene counts for each gene in each sample.
}
\description{
Use a peak-count matrix (rows=peaks, columns=samples) and as csAnno object from ChIPseeker
to aggregate counts from all or specific features for each gene
}
