% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrHeatMap.R
\name{chrHeatMap}
\alias{chrHeatMap}
\title{Plot a heatmap of binned signal across the whole genome}
\usage{
chrHeatMap(
  chrHeatMat,
  clusterSamples = TRUE,
  showColNames = TRUE,
  title = "",
  scaleValue = "score"
)
}
\arguments{
\item{chrHeatMat}{output of chrHeatMat() after binning. Not that smaller binSize may cause memory errors.}

\item{clusterSamples}{Boolean indicating if samples (columns) should be hierarchically clustered (TRUE) or remain in order (FALSE). Can be a data frame for grouping samples with row.names=names(GRList) and columns indicating annotations for groups.}

\item{showColNames}{Boolean indicating if sample names (column names) should be shown. Default is TRUE.}

\item{title}{character containing title for heatmap}

\item{scaleValue}{character containing the title for the heatmap scale- Default is 'score'}
}
\value{
Prints a ComplexHeatmap with binned chromosomes as rows, and samples as columns
}
\description{
Plot a heatmap of binned signal across the whole genome
}
