% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrHeatMap.R
\name{chrHeatMat}
\alias{chrHeatMat}
\title{Average signal across the genome in bins for a list of GRanges objects}
\usage{
chrHeatMat(GRList, scoreCol, binSize = 1e+05)
}
\arguments{
\item{GRList}{Named list of GRanges objects - each GRanges object should be from the same organism/genome build and should contain seqlengths}

\item{scoreCol}{Name of column in GRanges object containing the metric to be averaged across each bin}

\item{binSize}{Size of each genomic bin. Default is 100kb to help save memory for larger datasets}
}
\value{
a list containing a data frame ('mat') of the binned results and a GenomicRanges ('GR') object containing the binned results.
}
\description{
Takes a named list of GRanges objects containing a score column to be averaged in bins across the genome
}
