% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMER_Comparison.R
\name{compHOMER}
\alias{compHOMER}
\title{Function to compare two sets of HOMER known motif enrichment results
Results from HOMER known TF motif enrichment analyses are ranked either by enrichment or significance and plotted for comparison}
\usage{
compHOMER(
  xfile,
  yfile,
  title = "",
  xlab = "x",
  ylab = "y",
  lab = T,
  spec = NULL,
  numLab = 5,
  enrichment = F,
  returnRes = F
)
}
\arguments{
\item{xfile}{string of path to knownResults.txt file of a HOMER motif enrichment analysis}

\item{yfile}{string of path to knownResults.txt file of another HOMER motif enrichment analysis}

\item{title}{String indicating title of figure}

\item{xlab}{String indicating the x-axis label (describing xfile). Default="x"}

\item{ylab}{String indicating the y-axis label (describing yfile). Default="y"}

\item{lab}{Boolean indicating if certain TFs should be labeled}

\item{spec}{String vector indicating TFs of interest that should be labeled}

\item{numLab}{Numeric indicating the number of top ranked TFs to be labeled (default=5)}

\item{enrichment}{Boolean indicating if the results should be ranked by enrichment. FALSE ranks results by significance. Default=FALSE.}

\item{returnRes}{Boolean indicating if data frame of results should be returned. Default=F.}

\item{return}{Plot comparing HOMER results. Data frame of TFs and corresponding rankings and differences between those rankings.}
}
\description{
Function to compare two sets of HOMER known motif enrichment results
Results from HOMER known TF motif enrichment analyses are ranked either by enrichment or significance and plotted for comparison
}
