% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoFunctions.R
\name{getAnnoGenes}
\alias{getAnnoGenes}
\title{Subset and extract genes from ChIPseeker annotations}
\usage{
getAnnoGenes(anno, feature = "all", unique = T, geneCol = "SYMBOL")
}
\arguments{
\item{anno}{ChIPseeker annotation (csAnno) object}

\item{feature}{string indicating the genomic features to subset e.g. "promoter", "intergenic". Default is "all" and will not subset features.}

\item{unique}{Boolean indicating if only unique genes should be returned. Default is TRUE.}

\item{geneCol}{string indicating the column name in anno@anno containing the gene identifiers to be returned. Default is "SYMBOL" for gene symbols.}
}
\value{
Character vector of gene IDs.
}
\description{
Subset and extract genes from ChIPseeker annotations
}
