% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetagene.R
\name{makeCovList}
\alias{makeCovList}
\title{Function to import coverage matrix to R}
\usage{
makeCovList(
  matfile,
  header = 3,
  bedfile = NULL,
  sampleTable,
  annobed = F,
  species = "human",
  TxDb = NULL,
  db = NULL
)
}
\arguments{
\item{matfile}{Character indicating the path to the matrix file (ending in '.mat.gz')}

\item{header}{Numeric indicating the size of the header in matfile (i.e. number of lines to skip while reading file in). Default=3}

\item{bedfile}{Character indicating the path to the bed region files used in computeMatrix (can be left 'NULL' to use regions in .mat.gz file). MUST be same order as .mat.gz regions}

\item{sampleTable}{data.frame with two columns: sampleName and sampleID. sampleName contains the name that will appear for each sample in the figure. SampleID contains a pattern denoting each unique sample (can be the same as sampleName). NOTE: samples must be the same order as that of bigwigs input into computeMatrix.}

\item{annobed}{Boolean indicating if bed file should be annotated to include transcript IDs (default=FALSE)}

\item{species}{Character indicating species if annobed=T. Must be "human" or "mouse". Can set to "other" but TxDb object and annoDb must be provided.}

\item{TxDb}{TxDb object. Used if species="other", otherwise, "TxDb.Hsapiens.UCSC.hg38.knownGene" or "TxDb.Mmusculus.UCSC.mm10.knownGene" are used.}

\item{db}{Annotation database (as character) to use if species="other", otherwise "org.Hs.eg.db" or "org.Mm.eg.db" are used - must have corresponding bioconductor package installed.}
}
\value{
list of length 3: covList containing list of coverage matrices, bed containing regions for each row of the coverage matrices, and cols: a numeric vector of length 4 providing the suggested column parameters for plotMetaGene() and subsetquent functions. Use only when computeMatrix was executed with 'scale-regions' and NOT 'reference-point'
}
\description{
This function will import the .mat.gz coverage matrix generated by deeptools
computeMatrix command, and split it into a list of coverage matrices - one for
each bigwig file input into the command.
}
