% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetagene.R
\name{plotMetaGene}
\alias{plotMetaGene}
\title{Make a custom metagene plot}
\usage{
plotMetaGene(
  metaList,
  samples = NULL,
  subgenes = NULL,
  title = "",
  cols = NULL,
  marks = c("-500", "TSS", "TTS", "+2000"),
  xlab = "Position (bp)",
  ylab = "Occupancy",
  pal = "Dark2",
  lty = NULL,
  a = 100,
  b = 100,
  scale = F,
  ylim = NULL,
  ytick = NULL,
  retDat = F
)
}
\arguments{
\item{metaList}{metaList object output from makeCovList(), addSym(), or scaleMeta()}

\item{samples}{character vector indicating the names of the samples in metaList$cov to plot. NULL plots all.}

\item{subgenes}{Character vector of gene symbols to subset coverage matrices by. Must provide gene symbols to metaList after 'addSym()' has been used.}

\item{title}{Character indicating the title of the plot}

\item{cols}{Numeric vector of length 3 or 4. Length 3 indicates computeMatrix run with reference-point, length 4 indicates computeMatrix run with scale-regions. Leave NULL to use suggested cols paramters from metaList. Replace to override manually. This will affect any scaling (if indicated) and will affect the placement of the markers on the x-axis.}

\item{marks}{Character vector of length (length(cols)) indicating how landmarks should be labeled on the x-axis. Defaults to c("-500","TSS","TTS","+2000")}

\item{xlab}{Character for x-axis label. Default is "Position (bp)"}

\item{ylab}{Character for y-axis label. Default is "Occupancy".}

\item{pal}{Character indicating RColourBrewer palette name for colour schemes or custom character vector containing either colour names, hexadecimal or rgb() values.}

\item{lty}{Numeric indicating the line type for each sample. Default is NULL (lty=1 for each sample - solid line)}

\item{a}{numeric indicating the number of columns downstream region (cols[3]-cols[4]) should be scaled down to, if scaling indicated}

\item{b}{numeric indicating the number of columns upstream region (cols[1]-cols[2]) should be scaled down to, if scaling indicated}

\item{scale}{Boolean indicating if up/downstream regions should be scaled. Default is FALSE. If set to TRUE, must be on computeMatrix output from scale-regions, and cols must be length 4.}
                                                                              
\item{ylim}{Numeric vector of length 2 indicating min,max of y axis, respectively. Default is NULL for auto determination of limits.}                                                                     
                                                                              
\item{ytick}{Numeric vector indicating y-axis ticks. Leave NULL for default.}                                                                         

\item{retDat}{Indicating which type of data should be returned. Default NULL returns no data. Logical TRUE returns data for a metagene plot. "box" returns data for boxplots of coverage.}
}
\value{
Custom metagene figure
}
\description{
Make a custom metagene plot
}
