% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakVenn.R
\name{plotVenn_gr}
\alias{plotVenn_gr}
\title{Function to plot Venn diagram of overlapping peaks}
\usage{
plotVenn_gr(
  x,
  title = "Venn Diagram",
  cols = "Dark2",
  lty = "blank",
  scale = F,
  retVals = F
)
}
\arguments{
\item{x}{named list of GenomicRagnes objects to compare. Maximum 5 sets of regions.}

\item{title}{String indicating title of plot}

\item{cols}{colour palette to use. Can be RColorBrewer Palette, or string of rgb, hexadecimal, or colour names the same length of 'x'}

\item{lty}{Line type. Default="blank" for no lines. 1 for solid lines, 2 for dashed lines, etc.}

\item{scale}{Boolean indicating if circles should be scaled to the size of each region. Works for comaparisons of less than 4.}

\item{retVals}{Boolean indicating if overlap regions should be returned. Default=FALSE.}
}
\value{
Venn diagram plot and if retVals=T, a named list of GenomicRanges for each overlapping region
}
\description{
Function to plot Venn diagram of overlapping peaks
}
