% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoFunctions.R
\name{retRegion}
\alias{retRegion}
\title{Return genomic regions associated with a specific gene or set of genes}
\usage{
retRegion(anno, symbol)
}
\arguments{
\item{anno}{ChIPseeker annotation (csAnno) object}

\item{symbol}{String or vector of strings indicating the gene symbol(s) to search}
}
\value{
data frame containing all information about all peaks associated with gene(s) of interest
}
\description{
Return genomic regions associated with a specific gene or set of genes
}
