% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetagene.R
\name{scaleMeta}
\alias{scaleMeta}
\title{Apply the scalefun() to a metalist}
\usage{
scaleMeta(metaList, cols = c(0, 100, 300, 700), b = 100, a = 100)
}
\arguments{
\item{metaList}{output of either makeCovList() or addSym()}

\item{cols}{numeric vector of length 4 indicating landmark columns (start=0, end of upstream region, end of main region, ncol(df))}

\item{b}{numeric indicating the number of columns the upstream region (cols[1]-cols[2]) should be scaled down to. Default=100.}

\item{a}{numeric indicating the number of columns the downstream region (cols[3]-cols[4]) should be scaled down to. Default=100.}
}
\value{
metaList of coverage matrices with upstream and downstream regions scaled down (i.e. reduced columns) for clearer plotting
}
\description{
Apply the scalefun() to a metalist
}
