% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedFunctions.R
\name{writeBETA}
\alias{writeBETA}
\title{Function to export GenomicRanges object for use with Cistrome BETA}
\usage{
writeBETA(x, peakID, filename, scoreCol = "V7")
}
\arguments{
\item{x}{GenomicRanges object to export}

\item{peakID}{String or String vector to use to identify peak names (will be concatenated to the row number of each peak)}

\item{filename}{String of path to export results (Should end in .bed)}

\item{scorCol}{String indicating column name in GenomicRanges object that should be used for scoring peaks in BETA analysis. Default="V7"}
}
\description{
Function to export GenomicRanges object for use with Cistrome BETA
}
