% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedFunctions.R, R/FactorFootprints.R
\name{readBed}
\alias{readBed}
\title{Function to read tab-delimited bed file to GenomicRanges object}
\usage{
readBed(filename, HOCO = F, species = "human", BSg = NULL)

readBed(filename, HOCO = F, species = "human", BSg = NULL)
}
\arguments{
\item{filename}{Character of .bed file name and location (assumes no header)}

\item{HOCO}{Boolean indicaitng if .bed file is HOCOMOCO TFBS bed file. Default is FALSE}

\item{species}{One of "human" or "mouse" to use already specified BSgenome objects. Leave NULL to use custom BSgenome.}

\item{BSg}{BSgenome object containign seqlengths for genome of interest. Use if you want to specify specific BSgenome (especially if not mouse or human). Or leave NULL along with 'species' to not include seqlengths.}

\item{header}{Boolean indicating if there is a header in the bed file. Default=FALSE.}
}
\value{
GenomicRanges object

GRanges object of bedfile regions
}
\description{
Function to read tab-delimited bed file to GenomicRanges object

Read a bed file to genomicRanges Object with option for seqlengths
}
